使用LocalDateTime替换Date类型解决mysql表中createTime和updateTime快16小时的bug

+++ src/main/java/com/bitop/otcapi/fcg/entity/OtcCoinConfig.java
@@ -44,7 +44,7 @@ public class OtcCoinConfig extends BaseEntity implements Serializable {
     private double fee;
 
     @ApiModelProperty(value = "状态 （0启用 1禁用 ）")
-    private short status;
+    private int status;
 
     @ApiModelProperty(value = "停止销售（0启用 1禁用 ）")
     private int stopSelling;
